+++ src/sdk.ts
@@ -4,3 +4,4 @@ export * from "./manifest.js";
 export * from "./signing.js";
 export * from "./admit.js";
 export * from "./grant.js";
+export * from "./trust.js";